lg-trade/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.26
openai==1.97.0
pinecone==7.3.0
python-dotenv==1.1.1
langchain-community==0.3.27
langchain-openai==0.3.28
python-multipart==0.0.20
PyPDF2==3.0.1
langchain-pinecone==0.2.9
langgraph==0.5.3